+++ src/modules/tools/print/store/actions/actionsPrintInitialization.js
@@ -7,8 +7,8 @@ import getVisibleLayer from "./../../utils/getVisibleLayer";
 import {createEmpty, extendCoordinate} from "ol/extent.js";
 import {apply as applyTransform} from "ol/transform.js";
 import getProxyUrl from "../../../../../utils/getProxyUrl";
-import {autoDrawMask} from "olcs/print/drawCesiumMask.ts";
-import {computeRectangle} from "olcs/print/computeRectangle.ts";
+import {autoDrawMask} from "olcs/print/drawCesiumMask.js";
+import {computeRectangle} from "olcs/print/computeRectangle.js";
 
 let lastPrintedExtent;